Fix normalization of dates from different timezones
I noticed that timestamps from different timezones were parsed incorrectly. There were a number of small glitches in #normalizeDateAndTime:: Both DateAndTime >> #offset: and DateAndTime >> #utcOffset: are written in functional style, but their return value in the mailman aggregator was discarded. Also, make sure not to convert the time manually but instead to preserve the original offset. The originally intended behavior for this was not clear, but I see some value in saving this information since it could be relevant for further analysis. Callers that do not care about the offset can use DateAndTime >> #asLocal to convert the date into the local timezone.